feat: upd displaylink-debian kernel >= 6.6.0 (#920)
Starting with kernel version 6.6.0 and above the detection and configuration of
Intel graphic card subsystem does not work anymore.

With this patch i keep the behaviour as is for kernel <6.6.0 and
add a different one for newer kernels.

comment in displaylink-debian.sh
use different content in $cardsub to repair Intel Subsystem handling with kernel >= 6.6  (issue: 920)

Fixes AdnanHodzic/displaylink-debian#920